+++ atlas/data_structure.py
@@ -213,6 +213,10 @@ SITES_SCHEMA = {
             'cse_id': {
                 'type': 'string',
             },
+            'memcache': {
+                'type': 'boolean',
+                'default': True,
+            },
         },
     },
     'tag': {